+++ lib/src/constant/http_method.dart
@@ -1,0 +1,7 @@
+/// Method for request
+abstract class HttpMethod {
+  static const POST = 'POST';
+  static const GET = 'GET';
+  static const PUT = 'PUT';
+  static const DELETE = 'DELETE';
+}