+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 - publish beta version,and can use hosting to build a simple app
+
+## 0.0.2 
+
+- add a Method attribute to HttpRequest
+- add HttpStatus,HttpMethod,ContentType constant